+++ R/export.r
@@ -171,39 +171,157 @@ saveAsTxt <- function(x, file = NA) {
 ############################# EXPORT EXCEL ####################################
 
 
-#' Save grid as Microsoft Excel file (.xlsx)
+#' Save grids as Microsoft Excel file (.xlsx)
 #'
-#' `saveAsExcel` will save the grid as a Microsoft Excel file (`.xlsx`).
+#' `saveAsExcel` will save one or more grids as MS Excel file (`.xlsx`).
 #'
-#' @param x A `repgrid` object.
+#' @param x A `repgrid` object or a list of grids.
 #' @param file File path. Suffix must be `.xlsx`.
 #' @param format Two output formats are supported: `wide` (default) where each column represents one element, each row
 #'   represent one constructs (a common grid representation), and `long` where each row contains an element-construct
 #'   combination and the corresponding rating value. See [importExcel()] for details and examples.
-#' @param sheet  Optional sheet name. If `NULL` (default), either `"grid (wide format)"` or `"grid (long format)"`.
+#' @param sheet Sheet name (defaults to `grid`). If `x` is a list of grids, the name of the list entry is used.
+#'   If it has no name, a sequential index is appended to `sheet`.
 #' @return  Invisibly returns file path.
 #' @export
-#' @seealso  [importExcel()]
-#' @examples
-#' # save grid in wide format (default)
-#' file <- tempfile(fileext = ".xlsx")
-#' saveAsExcel(boeker, file)
-#' \dontrun{
-#' browseURL(file) # open file, requires Excel, may not work on all system
-#' }
+#' @seealso  [importExcel()], [saveAsWorksheet()]
+#' @example inst/examples/example-save-as-excel.R
 #'
-#' # save grid in log format
-#' file <- tempfile(fileext = ".xlsx")
-#' saveAsExcel(boeker, file, format = "long")
-#' \dontrun{
-#' browseURL(file)
-#' }
-saveAsExcel <- function(x, file, format = "wide", sheet = NULL) {
-  stop_if_not_is_repgrid(x)
+saveAsExcel <- function(x, file, format = "wide", sheet = "grid") {
   ext <- tools::file_ext(file)
   if (ext != "xlsx") {
     stop("The file extension must be '.xlsx'. Found '", ext, "' instead.", call. = FALSE)
   }
+  wb <- openxlsx::createWorkbook()
+  wb <- saveAsWorksheet(x, wb = wb, format = format, sheet = sheet)
+  openxlsx::saveWorkbook(wb, file, overwrite = TRUE)
+  invisible(file)
+}
+
+
+#' Add grids as sheets to an openxlsx Workbook
+#'
+#' `saveAsWorksheet` will add one or more grids to an a [openxlsx](https://CRAN.R-project.org/package=openxlsx)
+#' `Workbook` object.
+#'
+#' @param x A `repgrid` object or a list of grids.
+#' @param wb A [openxlsx](https://CRAN.R-project.org/package=openxlsx) `Workbook` object.
+#' @inheritParams saveAsExcel
+#' @return  Invisibly returns Workbook object.
+#' @export
+#' @seealso  [saveAsExcel()]
+#' @example inst/examples/example-save-as-worksheet.R
+#'
+saveAsWorksheet <- function(x, wb, format = "wide", sheet = "grid") {
+  stop_if_not_inherits(wb, "Workbook")
+
+  if (!is_list_of_repgrids(x)) {
+    wb <- add_one_sheet_with_grid(x, wb = wb, format = format, sheet = sheet)
+  } else {
+    ii <- seq_along(x)
+    list_names <- get_names_na(x)
+    sheets_numbered <- paste(sheet %||% "grid", ii)
+    sheets <- ifelse(!is.na(list_names), list_names, sheets_numbered)
+    for (i in ii) {
+      wb <- add_one_sheet_with_grid(x[[i]], wb = wb, format = format, sheet = sheets[i])
+    }
+  }
+  invisible(wb)
+}
+
+
+
+# saveAsExcelv2 <- function(x, file, format = "wide", sheet = "grid") {
+#   ext <- tools::file_ext(file)
+#   if (ext != "xlsx") {
+#     stop("The file extension must be '.xlsx'. Found '", ext, "' instead.", call. = FALSE)
+#   }
+#   if (is.null(wb)) {
+#     wb <- openxlsx::createWorkbook()
+#   }
+#   if (!is_list_of_repgrids(x)) {
+#     wb <- add_one_sheet_with_grid(x, wb = wb, format = format, sheet = sheet)
+#   } else {
+#     ii <- seq_along(x)
+#     list_names <- get_names_na(x)
+#     sheets_numbered <- paste(sheet %||% "grid", ii)
+#     sheets <- ifelse(is.na(list_names), sheets_numbered)
+#     for (i in ii) {
+#       wb <- add_one_sheet_with_grid(x[[i]], wb = wb, format = format, sheet = sheets[i])
+#     }
+#   }
+#   openxlsx::saveWorkbook(wb, file, overwrite = TRUE)
+#   invisible(file)
+# }
+#
+# saveAsExcel_v1 <- function(x, file, format = "wide", sheet = NULL) {
+#   if (is_list_of_repgrids(x)) {
+#
+#   }
+#
+#   stop_if_not_is_repgrid(x)
+#   ext <- tools::file_ext(file)
+#   if (ext != "xlsx") {
+#     stop("The file extension must be '.xlsx'. Found '", ext, "' instead.", call. = FALSE)
+#   }
+#   format <- match.arg(tolower(format), c("wide", "long"))
+#
+#   if (format == "wide") {
+#     df <- grid_to_df_wide(x)
+#     sheet <- sheet %||% "grid (wide format)"
+#     jj_rows <- seq_len(nrow(df) + 1)
+#     jj_elements <- seq_len(ncol(x)) + 1
+#     j_left_pole <- 1
+#     j_right_pole <- ncol(x) + 2
+#     j_preferred <- ncol(df)
+#
+#     style_header <- openxlsx::createStyle(textDecoration = "bold", fgFill = "grey90", border = "bottom")
+#     style_center <- openxlsx::createStyle(halign = "center")
+#     style_pole_left <- openxlsx::createStyle(textDecoration = "bold", halign = "right", border = "right")
+#     style_pole_right <- openxlsx::createStyle(textDecoration = "bold", halign = "left", border = "left")
+#     style_text_grey <- openxlsx::createStyle(fontColour = "grey50")
+#
+#     wb <- openxlsx::createWorkbook()
+#
+#     openxlsx::addWorksheet(wb, sheet)
+#     openxlsx::writeData(wb, sheet, x = df, headerStyle = style_header, rowNames = FALSE, colNames = TRUE)
+#     openxlsx::addStyle(wb, sheet, style = style_center, rows = jj_rows, cols = jj_elements, stack = TRUE, gridExpand = TRUE)
+#     openxlsx::addStyle(wb, sheet, style = style_pole_left, rows = jj_rows, cols = j_left_pole, stack = TRUE, gridExpand = TRUE)
+#     openxlsx::addStyle(wb, sheet, style = style_pole_right, rows = jj_rows, cols = j_right_pole, stack = TRUE, gridExpand = TRUE)
+#     openxlsx::addStyle(wb, sheet, style = style_text_grey, rows = jj_rows, cols = j_preferred, stack = TRUE, gridExpand = TRUE)
+#     openxlsx::setColWidths(wb, sheet, cols = c(j_left_pole, j_right_pole, j_preferred), widths = c(20, 20, 13))
+#     openxlsx::saveWorkbook(wb, file, overwrite = TRUE)
+#   } else {
+#     df <- grid_to_df_long(x)
+#     sheet <- sheet %||% "grid (long format)"
+#     jj_rows <- seq_len(nrow(df) + 1)
+#     jj_cols <- seq_len(ncol(df))
+#     jj_cols_optional <- 5:7
+#     j_preferred <- 5
+#     ii_end_construct <- rle(df$left_pole)$lengths %>% cumsum() %>% head(-1) + 1
+#
+#     style_header <- openxlsx::createStyle(textDecoration = "bold", fgFill = "grey90", border = "bottom")
+#     style_border_left <- openxlsx::createStyle(border = "left")
+#     style_border_bottom <- openxlsx::createStyle(border = "bottom", borderStyle = "dashed")
+#     style_text_grey <- openxlsx::createStyle(fontColour = "grey50")
+#
+#     wb <- openxlsx::createWorkbook()
+#     openxlsx::addWorksheet(wb, sheet)
+#     openxlsx::writeData(wb, sheet, x = df, headerStyle = style_header, rowNames = FALSE, colNames = TRUE)
+#     openxlsx::addStyle(wb, sheet, style = style_border_left, rows = jj_rows, cols = j_preferred, stack = TRUE, gridExpand = TRUE)
+#     openxlsx::addStyle(wb, sheet, style = style_text_grey, rows = jj_rows, cols = jj_cols_optional, stack = TRUE, gridExpand = TRUE)
+#     openxlsx::addStyle(wb, sheet, style = style_border_bottom, rows = ii_end_construct, cols = jj_cols, stack = TRUE, gridExpand = TRUE)
+#     openxlsx::setColWidths(wb, sheet, cols = jj_cols, widths = c(20, 20, 20, 10, 13, 10, 10))
+#     openxlsx::freezePane(wb, sheet, firstActiveRow = 2)
+#     openxlsx::saveWorkbook(wb, file, overwrite = TRUE)
+#   }
+#   invisible(file)
+# }
+
+add_one_sheet_with_grid <- function(x, wb, format = "wide", sheet = NULL) {
+  stop_if_not_is_repgrid(x)
+  stop_if_not_inherits(wb, "Workbook")
+
   format <- match.arg(tolower(format), c("wide", "long"))
 
   if (format == "wide") {
@@ -221,8 +339,6 @@ saveAsExcel <- function(x, file, format = "wide", sheet = NULL) {
     style_pole_right <- openxlsx::createStyle(textDecoration = "bold", halign = "left", border = "left")
     style_text_grey <- openxlsx::createStyle(fontColour = "grey50")
 
-    wb <- openxlsx::createWorkbook()
-
     openxlsx::addWorksheet(wb, sheet)
     openxlsx::writeData(wb, sheet, x = df, headerStyle = style_header, rowNames = FALSE, colNames = TRUE)
     openxlsx::addStyle(wb, sheet, style = style_center, rows = jj_rows, cols = jj_elements, stack = TRUE, gridExpand = TRUE)
@@ -230,7 +346,6 @@ saveAsExcel <- function(x, file, format = "wide", sheet = NULL) {
     openxlsx::addStyle(wb, sheet, style = style_pole_right, rows = jj_rows, cols = j_right_pole, stack = TRUE, gridExpand = TRUE)
     openxlsx::addStyle(wb, sheet, style = style_text_grey, rows = jj_rows, cols = j_preferred, stack = TRUE, gridExpand = TRUE)
     openxlsx::setColWidths(wb, sheet, cols = c(j_left_pole, j_right_pole, j_preferred), widths = c(20, 20, 13))
-    openxlsx::saveWorkbook(wb, file, overwrite = TRUE)
   } else {
     df <- grid_to_df_long(x)
     sheet <- sheet %||% "grid (long format)"
@@ -245,8 +360,6 @@ saveAsExcel <- function(x, file, format = "wide", sheet = NULL) {
     style_border_bottom <- openxlsx::createStyle(border = "bottom", borderStyle = "dashed")
     style_text_grey <- openxlsx::createStyle(fontColour = "grey50")
 
-    wb <- openxlsx::createWorkbook()
-    sheet <- "grid (long format)"
     openxlsx::addWorksheet(wb, sheet)
     openxlsx::writeData(wb, sheet, x = df, headerStyle = style_header, rowNames = FALSE, colNames = TRUE)
     openxlsx::addStyle(wb, sheet, style = style_border_left, rows = jj_rows, cols = j_preferred, stack = TRUE, gridExpand = TRUE)
@@ -254,9 +367,8 @@ saveAsExcel <- function(x, file, format = "wide", sheet = NULL) {
     openxlsx::addStyle(wb, sheet, style = style_border_bottom, rows = ii_end_construct, cols = jj_cols, stack = TRUE, gridExpand = TRUE)
     openxlsx::setColWidths(wb, sheet, cols = jj_cols, widths = c(20, 20, 20, 10, 13, 10, 10))
     openxlsx::freezePane(wb, sheet, firstActiveRow = 2)
-    openxlsx::saveWorkbook(wb, file, overwrite = TRUE)
   }
-  invisible(file)
+  wb
 }